Z-Wave: Node Deletion & Add Controller IsFailedNode
A Z-Wave node can only be deleted if its present in the controllers
"Failed List" resident in its own NVRAM.
Added the Z-Wave IsFailedNode Message Class.
Added new Failed status if the controller has it flagged as failed in
its own database.
This then allows a node to be deleted.
Visibility of the "Delete Node" button uses this new logic.

+++ bundles/binding/org.openhab.binding.zwave/src/main/java/org/openhab/binding/zwave/internal/config/ZWaveConfiguration.java
@@ -35,6 +35,7 @@ import org.openhab.binding.zwave.internal.protocol.commandclass.ZWaveVersionComm
 import org.openhab.binding.zwave.internal.protocol.commandclass.ZWaveWakeUpCommandClass;
 import org.openhab.binding.zwave.internal.protocol.event.ZWaveEvent;
 import org.openhab.binding.zwave.internal.protocol.event.ZWaveInclusionEvent;
+import org.openhab.binding.zwave.internal.protocol.event.ZWaveNetworkEvent;
 import org.openhab.binding.zwave.internal.protocol.initialization.ZWaveNodeSerializer;
 import org.osgi.framework.FrameworkUtil;
 import org.slf4j.Logger;
@@ -273,7 +274,7 @@ public class ZWaveConfiguration implements OpenHABConfigurationService, ZWaveEve
 				// Set the state
 				boolean canDelete = false;
 				switch(node.getNodeStage()) {
-				case DEAD:
+				case FAILED:
 					record.state = OpenHABConfigurationRecord.STATE.ERROR;
 					canDelete = true;
 					break;
@@ -289,10 +290,11 @@ public class ZWaveConfiguration implements OpenHABConfigurationService, ZWaveEve
 						record.state = OpenHABConfigurationRecord.STATE.WARNING;
 					else
 					record.state = OpenHABConfigurationRecord.STATE.OK;
+					canDelete = false;
 					break;
 				default:
 					record.state = OpenHABConfigurationRecord.STATE.INITIALIZING;
-					canDelete = true;
+					canDelete = false;
 					break;
 				}
 
@@ -830,15 +832,17 @@ public class ZWaveConfiguration implements OpenHABConfigurationService, ZWaveEve
 
 				if (action.equals("Delete")) {
 					logger.debug("NODE {}: Delete node", nodeId);
-					this.zController.requestRemoveFailedNode(nodeId);
+					if (node.isFailed()) {
+						this.zController.requestRemoveFailedNode(nodeId);
+						// Delete the XML file.
+						// TODO: This should be possibly be done after registering
+						// an event handler
+						// Then we can delete this after the controller confirms the
+						// removal.
+						ZWaveNodeSerializer nodeSerializer = new ZWaveNodeSerializer();
+						nodeSerializer.DeleteNode(nodeId);
+					}
 
-					// Delete the XML file.
-					// TODO: This should be possibly be done after registering
-					// an event handler
-					// Then we can delete this after the controller confirms the
-					// removal.
-					ZWaveNodeSerializer nodeSerializer = new ZWaveNodeSerializer();
-					nodeSerializer.DeleteNode(nodeId);
 				}
 
 				// This is temporary
